Rework scriptlet printing and callbacks
This adds more info to the scriptlet callbacks and adds callbacks for
when and install file is about to run/finished running.

Indent and colour the scriptlet output so the user can see it clearly.

Also pass the errno if execv fails back to the parent so it can do the
logging instead of it being part of the childs's output.